use chezmoi script dir, remove bloat

+++ .chezmoiscripts/linux/debian/run_onchange_before_bootstrap-wsl.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
-echo "💡 Enabling OpenSSH on WSL..."
+if [ "$INSTALLATION_TYPE" != "wsl" ]; then
+  echo "🚫 This script is only intended to run on WSL."
+  exit 1
+fi
+
+echo "🔧 Enabling OpenSSH on WSL..."
 sudo apt remove -y --purge openssh-server
 sudo apt install -y openssh-server
 sudo sed -i 's/#Port 22/Port 2222/g' /etc/ssh/sshd_config
@@ -14,5 +19,5 @@ sudo grep -qxF '[boot]' /etc/wsl.conf || echo '[boot]' | sudo tee -a /etc/wsl.co
 sudo grep -qxF 'command = service ssh start' /etc/wsl.conf || echo 'command = service ssh start' | sudo tee -a /etc/wsl.conf
 
 # you can access the SSH server through proxyjump
-echo "💡 Enabling piping of 1password SSH agent to WSL..."
+echo "🔐 Enabling piping of 1password SSH agent to WSL..."
 sudo apt install -y socat
